+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 _Bug fixes_
 
-* Paused nested pipelines now correctly pause parent pipelines. ([#955](https://github.com/turbot/flowpipe/issues/955)).
+* Nested pipelines now correctly pauses parent pipelines if required. ([#955](https://github.com/turbot/flowpipe/issues/955)).
+* Pipeline with `max_concurrency` setting is now automatically paused and will successfully resume. ([#957](https://github.com/turbot/flowpipe/issues/957)).
 * Fix crashing cases when using `--output json`. ([#594](https://github.com/turbot/pipe-fittings/issues/594)).
 * Coerce variables set in interactive console to their declared type. ([#595](https://github.com/turbot/pipe-fittings/issues/595)).
 * `form_url` is now sanitized.